backend/requirements.txt
flask
flask-cors
pymongo
razorpay
python-dotenv
werkzeug